#12: Create pin view model and basic view
This will be the foundation for which the pin detail view will operate.
The view model handles updating the pin at the correct location without
any awkward searching for the element, and the view updates in real
time, propagating back to the file.

On macOS, this is currently configured to display as an inspector on the
main window. This might change depending on UX testing.