#12: Allow selection from files on macOS
Users might want to pick screenshots from their hard disk on a Mac,
rather than a photo from the photo picker. On macOS, the Add Photos
button now provides a dropdown menu that lets the user choose between
the Photos picker and the regular file selection flow. The latter opens
an NSOpenPanel to open an image, and the upload process works the same.