#1: Display map ornaments on iOS, navigator wheel
To move around the map, a new directional navigator wheel has been
introduced. This allows players to jump between map chunks currently.

Additionally, for iOS, this counts as an ornament. Ornaments are now
properly displayed on the map, and performing a single tap will show or
hide them. Other ornaments include the location badge and the dimension
picker (only on iOS).