#10: Create document launch view
Moves the docuument launch view into its own structure, with macOS and
iOS variants. The Mac version matches more closely to the mockup and has
options for creating and opening maps, along with a recents list. There
is also a new Window item to bring up the launch experience on macOS.

Fixes #10.