Declare required encryption information
In App Store Connect, we're required to declare whether Alidade uses
encryption algorithms that would not be exempt from specific
regulations. Since Alidade only uses what's provided on Apple's
platforms out of the box, this can be easily declared as such in the
Info.plist file, thereby speeding up the process for deploying
TestFlight builds and prepping builds for the App Store.