Start implementing file versioning
The VersionedCodable package will help migrate versioned types over from
one iteration to the next, allowing a more flexible system for updating
the MCMap package format. Currently, v1 is the only type available, but
this will be updated as development progresses.

Tests should be written when the new version is introduced to handle the
migration aspects.