Create InlineBanner component
The inline banner view is used to display messaging inside of views,
such as on the map creator form. The colors adapt based on the color
scheme, and the accessibility readouts have been updated to also read
out the banner's state.

Likewise, the map creator form has been refactored to leverage this new
view.