Reorganize views, standardize routing
Routing in Red Window now is handled at the app level, and it can be
accessed through an environment value. This also allows for integrations
with the menu bar.

The project's files have also been reorganized to make it clearer which
code is shared, which is legacy, and which is Red Window.