Refactor pin detail view, animate navigation
The detail view is now broken up into individual cells similar to a
collection view. This makes the detail view a little more modular, and
it improves type checking performance.

Moreover, if a pin has a set of images available, the first image in the
list will be used as a cover in both the grid cell and detail header.

Finally, when tapping on a pinned place, a zooming animation is used
similar to Apple Music instead of a generic page push. This makes the
interaction feel more dynamic and grounded.