Clean up tests that fail on iOS 26
I'm not sure whether this is a beta issue or the test _do_ need to be
rewritten to account for SwiftUI changes, but most of these are just
wrapped with 'withKnownIssue' for now until I see how it plays out in
further releases. Likewise, the tests for the document view model have
been altered to only run on macOS, as it's most prevalent there
(previously, it would fail because it would try to work with URLs from
the simulator).