Add Hashable conformance to CartographyMapFile
This should allow passing the file through multiple windows, and reduce
the overhead on certain contexts. The only major concern here would be
the app state, as the JSON encoder might fail when trying to encode the
tab customizations. However, I suspect this would be rare unless there's
a bug in either the encoder or the TabViewCustomization type, both of
which are out of my control.