Migrate to CartographyMapPin
The future of the format will store pins in a separate location instead
of as part of the manifest. This mega-commit updates the app to point to
both the new type and location, respectively.

Incidentally, this fixes #45 because the pin route under Red Window now
has a stable identity that doesn't get edited magically.