Allow specifying pin dimensions
Some pins might only be applicable to certain dimensions such as the
Nether. In these cases, players can set the pin's dimension to have it
only appear in that dimension. The coordinates then are relative to the
dimension they're assigned to, and the appropriate label views will
automatically convert when necessary.

By default, pins in v1 will be assumed as the Overworld.